# Prefix trimming