Hello @sindresorhus. I just implemented the prefix trimming optimization on your function. However, on my benchmark, the benefit is not that clear cut (benchmark's examples may not be perfectly suited to see the advantages of this optimization maybe).

So I'll let you decide whether you want to merge this or not.

Have a good day.