@sindresorhus 2017-02-17
Nice!

Benchmarks looks good (Node.js 7.5.0).


## master


```
         401,487 op/s » leven
         371,707 op/s » talisman
         264,191 op/s » levenshtein-edit-distance
         152,923 op/s » fast-levenshtein
          57,267 op/s » levenshtein-component
          19,915 op/s » levdist
          21,802 op/s » ld
          18,079 op/s » natural
          11,761 op/s » levenshtein
```


## This PR

```
         448,206 op/s » leven
         362,329 op/s » talisman
         276,395 op/s » levenshtein-edit-distance
         146,641 op/s » fast-levenshtein
          61,061 op/s » levenshtein-component
          20,720 op/s » levdist
          22,642 op/s » ld
          18,966 op/s » natural
          11,623 op/s » levenshtein
```

Although `talisman` seems to have regressed slightly. My fixtures/benchmarks might not be optimal. Help welcome to make them more realistic.