@Yomguithereal 2017-02-17
How many times did you run the benchmarks? On my computer, they tend to differ a lot sometimes.