@sindresorhus 2017-02-17
Multiple times, but now I get slightly different results, so results are not very stable.

```
         439,409 op/s » leven
         384,934 op/s » talisman
         284,955 op/s » levenshtein-edit-distance
         161,983 op/s » fast-levenshtein
          59,220 op/s » levenshtein-component
          20,824 op/s » levdist
          23,470 op/s » ld
          20,145 op/s » natural
          12,604 op/s » levenshtein
```