@Yomguithereal 2017-02-17
Yes. I guess this is the fun of benchmarking JavaScript :).Do you know if `matcha` interleaves the tests and/or run them in separate threads each time?

Another solution to compare would be to copy the old `leven` function and incorporate it to the benchmark. What do you think?