@Yomguithereal 2017-02-17
I've pushed a branch [here](https://github.com/sindresorhus/leven/tree/benchmark-old-leven) to play with the idea.