@Yomguithereal 2017-02-17
And this:

```js
fn('bbbbbbackkkkkk', 'bbbbbbookkkkkk');
```

is a caricatural benchmark case if you want to see benefits of the trimming optimizations.