@sindresorhus 2017-02-18
> Do you know if matcha interleaves the tests and/or run them in separate threads each time?

It runs them serially in the same process. We can try increasing `set('mintime', 500);` and `set('delay', 100);` and see if it helps.