@sindresorhus 2017-02-18
> Another solution to compare would be to copy the old leven function and incorporate it to the benchmark. What do you think?

Yes, that's a good idea. We can just depend on the previous version through a git tag: `npm i --save-dev sindresorhus/leven#v2.0.0`