@sindresorhus 2017-02-18
This looks great. Thank you! :)